Rename user.py to users.py for consistency with route names
